fix(engine): limit expire msg commands in result
This adds a virtual limit to the number of EXPIRE Message commands the
MessageTimeToLiveChecker adds to a result when executed, to prevent that
it clogs the log stream with too many of these commands.

It achieves this by keeping track of the current visitor index, and
reusing this index in subsequent executions. This utilizes the new
column family method whileTrue starting at key in the state class. Note
that the index is a combination of the deadline and the message key, as
these make up the key of the column family.

This index must be reset once the iteration reaches the point that there
are no messages with an expired deadline in the column family any more.
The reset is necessary to make sure we don't skip any messages whose
deadline has expired, but was added during or after the iteration.

Determining whether or not to reset the index is difficult because the
visit method (and the underlying whileTrue) do not inform us that the
end was reached. To help readers understand the code, I've described
this in a comment block in the code.

The limit is not yet configurable, because this requires changes to the
`broker` module, i.e., exposing the config to the engine. Since this
module is shared between ZPA and ZDP, I prefer to do this in a separate
pull request, to make that easier to review.

Note that this commit breaks userspace, because the engine no longer
favors message expirations over regular processing and instead provides
a more equal balance between the two. It's unlikely that users depend on
this behavior.