test(engine): coverage for user task due and follow-up dates
Add test coverage for the new dueDate and followUpDate properties on
user tasks. These tests cover user task activation. They validate that
the new properties are attached to the JobIntent.CREATED event.

The tests also assert that incidents are raised if the expressions for
dueDate and followUpDate can't be properly evaluated due to a missing
variable, invalid date format, or invalid expression result type.

It should be noted that when a dueDate or followUpDate expression
outputs and empty String, an incident is raised, as this is an invalid
date format. This differs from the behavior of other user task
properties like assignee, candidateUsers, or candidateGroups. The reason
is that an empty string can be seen as a valid user, or empty candidate
list and this can be easily ignored. This is not the case for dates.